Unwanted deletion
Ajout  d'une question en cas de save pour demander si l'utilisateur veut supprimer les fichiers précédents.

+++ Variables.py
@@ -27,7 +27,7 @@ mode : choisis le mode d'intégration. Trois possibilités :
 '''
 mode = 'iterations'
    # Combien de timesteps au maximum on s'autorise
-nitermax = 5000
+nitermax = 10000
    # Nombre de pixels du domaine :
 NX = 256*1 ; NY = 128*1 + 1
    # Largeur du canal. On doit avoir LY > 2*L
@@ -138,11 +138,10 @@ l_objects = []
 #Exemple de création d'une barre tournant de manière constante
 barre = R_Bar(long=2,larg=1,x_c=x_c,y_c=y_c)
 barre.period = 50 #Periode de rotation 50s
-barre.set_function(constant_rotation)
 
-barre2 = R_Bar(x_c = x_c,y_c=y_c,long=2,larg=0.2,period=50)
+barre2 = R_Bar(x_c = x_c+3,y_c=y_c+3,long=2,larg=0.5)
 barre2.set_function(apply_pressure)
-
+l_objects.append(Cylinder(x_c=x_c,y_c=y_c))
 #l_objects.append(barre)
 l_objects.append(barre2)
 
@@ -179,8 +178,9 @@ if (save):
         else :
             print('qualite_video n\'est pas defini correctement')
     files = os.listdir(save_directory)
-    for name in files:
-        os.remove(os.path.join(save_directory,name)) # supprime tout ce qu'il y avait avant
+    if input("Delete existing files? [y/n]")=="y":
+        for name in files:
+            os.remove(os.path.join(save_directory,name)) # supprime tout ce qu'il y avait avant
    # la figure est actualisée tous les 'modulo' timesteps
 modulo = nitermax//nFrames 
 if (mode == 'time'):
